worker/requirements.txt
boto3==1.42.78
botocore==1.42.78
certifi==2026.2.25
charset-normalizer==3.4.7
idna==3.11
jmespath==1.1.0
kafka-python==2.3.0
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
urllib3==1.26.20
